+++ src/main/java/com/example/demo/dto/MateriaGrupoDto.java
@@ -1,13 +1,11 @@
 package com.example.demo.dto;
 
-import com.example.demo.repositories.HorarioRepository;
-
 import java.util.List;
 
 public class MateriaGrupoDto {
 
-    private Integer grupoId;
-    private Integer materiaId;
+    private Long grupoId;
+    private Long materiaId;
     private String modalidad;
     private Integer numeroGrupo;
     private String aula;
@@ -16,9 +14,8 @@ public class MateriaGrupoDto {
     private Integer creditos;
     private String descripcion;
 
-    
-    public MateriaGrupoDto(Integer grupoId, Integer materiaId, String modalidad, Integer numeroGrupo, String aula,
-                           List<HorarioDto> horario, String nombre, Integer creditos, String descripcion) {
+    public MateriaGrupoDto(Long grupoId, Long materiaId, String modalidad, Integer numeroGrupo, String aula,
+            List<HorarioDto> horario, String nombre, Integer creditos, String descripcion) {
         this.grupoId = grupoId;
         this.materiaId = materiaId;
         this.modalidad = modalidad;
@@ -30,51 +27,66 @@ public class MateriaGrupoDto {
         this.descripcion = descripcion;
     }
 
-    public Integer getGrupoId() {
+    public Long getGrupoId() {
         return grupoId;
     }
-    public void setGrupoId(Integer grupoId) {
+
+    public void setGrupoId(Long grupoId) {
         this.grupoId = grupoId;
     }
-    public Integer getMateriaId() {
+
+    public Long getMateriaId() {
         return materiaId;
     }
-    public void setMateriaId(Integer materiaId) {
+
+    public void setMateriaId(Long materiaId) {
         this.materiaId = materiaId;
     }
+
     public String getModalidad() {
         return modalidad;
     }
+
     public void setModalidad(String modalidad) {
         this.modalidad = modalidad;
     }
+
     public Integer getNumeroGrupo() {
         return numeroGrupo;
     }
+
     public void setNumeroGrupo(Integer numeroGrupo) {
         this.numeroGrupo = numeroGrupo;
     }
+
     public String getAula() {
         return aula;
     }
+
     public void setAula(String aula) {
         this.aula = aula;
     }
+
     public String getNombre() {
         return nombre;
     }
+
     public void setNombre(String nombre) {
         this.nombre = nombre;
     }
+
     public Integer getCreditos() {
         return creditos;
     }
+
     public void setCreditos(Integer creditos) {
         this.creditos = creditos;
     }
+
     public String getDescripcion() {
         return descripcion;
     }
+
     public void setDescripcion(String descripcion) {
         this.descripcion = descripcion;
     }
