+++ src/main/java/com/example/demo/repositories/HorarioEstRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HorarioEstRepository extends CrudRepository<HorarioEstModel, Integer>{
+public interface HorarioEstRepository extends CrudRepository<HorarioEstModel, Long>{
     
 }
